feature: allow sites to be disabled through file naming
A small change that allows a site file configuration to be disabled by ending it with `.disabled.js`.

I've found this easier/preferable to temporarily deleting the file.

Another approach would be to add a `enabled: true/false` property, but this was simpler to implement -- and I think  it's more obvious if a site is disabled.

Maybe you'll like the idea.